time_series_agl.R
Uses NON-Reversed .csv files as the inputs and can only take in the 5 series that we selected. Calculates forecasts for 5 days into the future for each of the five stocks and outputs them to a file named History.csv, which will be created in your R workspace directory.

The history file is formatted such that the oldest 25 (5 days for each 5 stocks) predictions are listed at the top, to the newest 25 predictions at the bottom of the .csv file. This contains 6 columns: Point estimate, 80 CI Low, 80 CI High, 95 CI Low, 95 CI High, and the series tag i.e. BCH, BTC, etc.

Notes:
*Number of historical iterations is currently set at 10. You may change this at the top of the file.
*Expect this calculation to take a while